+++ packages/client/src/actions/applyAction.ts
@@ -30,7 +30,7 @@ import type {
   DataValue,
   SyncApplyActionResponseV2,
 } from "@osdk/foundry.ontologies";
-import * as OntologiesV2 from "@osdk/foundry.ontologies";
+import * as Actions from "@osdk/foundry.ontologies/Action";
 import type { MinimalClient } from "../MinimalClientContext.js";
 import { addUserAgentAndRequestContextHeaders } from "../util/addUserAgentAndRequestContextHeaders.js";
 import { augmentRequestContext } from "../util/augmentRequestContext.js";
@@ -129,7 +129,7 @@ export async function applyAction<
     action,
   );
   if (Array.isArray(parameters)) {
-    const response = await OntologiesV2.Actions.applyBatch(
+    const response = await Actions.applyBatch(
       clientWithHeaders,
       await client.ontologyRid,
       action.apiName,
@@ -153,7 +153,7 @@ export async function applyAction<
       ? edits?.type === "edits" ? remapActionResponse(response) : edits
       : undefined) as ActionReturnTypeForOptions<Op>;
   } else {
-    const response = await OntologiesV2.Actions.apply(
+    const response = await Actions.apply(
       clientWithHeaders,
       await client.ontologyRid,
       action.apiName,