+++ packages/client/src/observable/internal/BulkObjectLoader.ts
@@ -20,7 +20,6 @@ import type {
   ObjectTypeDefinition,
 } from "@osdk/api";
 import { PalantirApiError } from "@osdk/shared.net.errors";
-import { DefaultMap, DefaultWeakMap } from "mnemonist";
 import type { DeferredPromise } from "p-defer";
 import pDefer from "p-defer";
 import { additionalContext, type Client } from "../../Client.js";
@@ -28,6 +27,8 @@ import type {
   ObjectHolder,
 } from "../../object/convertWireToOsdkObjects/ObjectHolder.js";
 import type { DefType } from "../../util/interfaceUtils.js";
+import { DefaultMap } from "./collections/DefaultMap.js";
+import { DefaultWeakMap } from "./collections/DefaultWeakMap.js";
 
 interface InternalValue {
   primaryKey: string;