+++ packages/react-components/src/object-table/TableBody.tsx
@@ -32,6 +32,9 @@ interface TableBodyProps<TData extends RowData> {
   ) => React.ReactNode;
   isLoadingMore?: boolean;
   headerGroups?: Array<HeaderGroup<TData>>;
+  focusedRowId?: string | null;
+  setFocusedRowId?: (rowId: string | null) => void;
+  isInEditMode?: boolean;
 }
 
 export function TableBody<TData extends RowData>({
@@ -42,6 +45,9 @@ export function TableBody<TData extends RowData>({
   rowHeight = 40,
   isLoadingMore = false,
   headerGroups = [],
+  focusedRowId,
+  setFocusedRowId,
+  isInEditMode,
 }: TableBodyProps<TData>): React.ReactElement {
   // Important: Keep the row virtualizer in the lowest component possible to avoid unnecessary re-renders.
   const rowVirtualizer = useVirtualizer<HTMLDivElement, HTMLTableRowElement>({
@@ -80,6 +86,9 @@ export function TableBody<TData extends RowData>({
             virtualRow={virtualRow}
             onRowClick={onRowClick}
             renderCellContextMenu={renderCellContextMenu}
+            focusedRowId={focusedRowId}
+            setFocusedRowId={setFocusedRowId}
+            isInEditMode={isInEditMode}
           />
         );
       })}