+++ packages/react-components/src/object-table/hooks/useObjectTableData.ts
@@ -18,7 +18,6 @@ import type {
   DerivedProperty,
   ObjectOrInterfaceDefinition,
   ObjectSet,
-  ObjectTypeDefinition,
   PropertyKeys,
   QueryDefinition,
   SimplePropertyDef,
@@ -115,9 +114,9 @@ export function useObjectTableData<
   const objectSetResult = useObjectSet(
     shouldUseObjectSet ? objectSet as ObjectSet<Q> : undefined as any,
     {
-      ...(objectSetOptions as ObjectSetOptions<ObjectTypeDefinition>),
+      ...(objectSetOptions as ObjectSetOptions<Q>),
       withProperties: withProperties as WithProperties<
-        ObjectTypeDefinition,
+        Q,
         RDPs
       >,
       where: filter,
@@ -143,10 +142,6 @@ export function useObjectTableData<
 
   // Return the appropriate result based on which hook is enabled
   if (shouldUseObjectSet) {
-    // eslint-disable-next-line no-console
-    console.debug(
-      "Using useObjectSet for data fetching in ObjectTable because objectSet is provided and objectType is an object.",
-    );
     // Convert UseObjectSetResult to UseOsdkListResult format
     return {
       data: objectSetResult.data,
@@ -158,9 +153,5 @@ export function useObjectTableData<
     } as UseOsdkListResult<Q, RDPs>;
   }
 
-  // eslint-disable-next-line no-console
-  console.debug(
-    "Using useOsdkObjects for data fetching in ObjectTable because objectSet is not provided or objectType is an interface.",
-  );
   return osdkObjectsResult;
 }