+++ packages/create-app.template.tutorial-todo-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @osdk/create-app.template.tutorial-todo-app
 
+## 2.8.0-beta.15
+
+## 2.8.0-beta.14
+
+### Minor Changes
+
+- 35f2f1a: Add Media inputs/outputs for Queries
+- 6260835: Revert node bump
+
 ## 2.8.0-beta.13
 
 ## 2.8.0-beta.12