+++ packages/foundry-config-json/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/foundry-config-json
 
+## 1.7.0-beta.2
+
+### Minor Changes
+
+- 35f2f1a: Add Media inputs/outputs for Queries
+
 ## 1.4.0
 
 ### Patch Changes