+++ packages/generator-converters.ontologyir/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @osdk/generator-converters.ontologyir
 
+## 2.8.0-beta.15
+
+### Patch Changes
+
+- @osdk/client.unstable@2.8.0-beta.15
+
+## 2.8.0-beta.14
+
+### Minor Changes
+
+- 35f2f1a: Add Media inputs/outputs for Queries
+
+### Patch Changes
+
+- Updated dependencies [35f2f1a]
+  - @osdk/client.unstable@2.8.0-beta.14
+
 ## 2.8.0-beta.13
 
 ### Patch Changes