+++ packages/react-components-styles/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/react-components-styles
 
+## 0.2.0-beta.8
+
+### Minor Changes
+
+- 447e29d: Update editable table UX and API to pass the entire row data on edit
+
+## 0.2.0-beta.7
+
+### Minor Changes
+
+- 35f2f1a: Add Media inputs/outputs for Queries
+
 ## 0.2.0-beta.6
 
 ### Minor Changes