+++ packages/shared.net.fetch/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @osdk/shared.net.fetch
 
+## 1.8.0-beta.3
+
+### Minor Changes
+
+- 4a79eb1: Support QoS-Retry-Hint response header
+
+## 1.8.0-beta.2
+
+### Minor Changes
+
+- 35f2f1a: Add Media inputs/outputs for Queries
+
+### Patch Changes
+
+- Updated dependencies [35f2f1a]
+  - @osdk/shared.net.errors@2.8.0-beta.2
+
 ## 1.5.0
 
 ### Patch Changes